+++ FCL/src/main/java/com/mio/manager/RendererManager.kt
@@ -3,8 +3,8 @@ package com.mio.manager
 import android.content.Context
 import com.mio.data.Renderer
 import com.tungsten.fcl.R
-import com.tungsten.fclauncher.plugins.DriverPlugin
-import com.tungsten.fclauncher.plugins.RendererPlugin
+import com.mio.plugin.DriverPlugin
+import com.mio.plugin.RendererPlugin
 import com.tungsten.fcl.FCLApp
 
 object RendererManager {
