+++ FCL/src/main/java/com/mio/plugin/DriverPlugin.kt
@@ -1,7 +1,6 @@
-package com.tungsten.fclauncher.plugins
+package com.mio.plugin
 
 import android.content.Context
-import com.mio.manager.PluginManager
 import com.tungsten.fcl.FCLApp
 
 object DriverPlugin {
