+++ FCL/src/main/java/com/mio/ui/dialog/DriverSelectDialog.kt
@@ -7,7 +7,7 @@ import android.widget.ArrayAdapter
 import com.tungsten.fcl.R
 import com.tungsten.fcl.databinding.DialogSelectRendererBinding
 import com.tungsten.fcl.setting.Profiles
-import com.tungsten.fclauncher.plugins.DriverPlugin
+import com.mio.plugin.DriverPlugin
 import com.tungsten.fcllibrary.component.dialog.FCLDialog
 import com.tungsten.fcllibrary.util.ConvertUtils
 import java.util.function.Consumer
